+++ examples/camera/test.py
@@ -7,6 +7,7 @@ import time
 
 g_results = None
 g_normalized_images = []
+index = 0
 
 def callback(results):
     global g_results
@@ -18,6 +19,7 @@ def showNormalizedImage(name, normalized_image):
     return mat
     
 def process_video(scanner):
+    global g_normalized_images, index
     scanner.addAsyncListener(callback)
     
     cap = cv2.VideoCapture(0)
@@ -29,29 +31,37 @@ def process_video(scanner):
             break
         elif ch == ord('n'): # normalize image
             if g_results != None:
+                
+                if len(g_results) > 0:
+                    for result in g_results:
+                        x1 = result.x1
+                        y1 = result.y1
+                        x2 = result.x2
+                        y2 = result.y2
+                        x3 = result.x3
+                        y3 = result.y3
+                        x4 = result.x4
+                        y4 = result.y4
+                        
+                        normalized_image = scanner.normalizeBuffer(image, x1, y1, x2, y2, x3, y3, x4, y4)
+                        g_normalized_images.append((str(index), normalized_image))
+                        showNormalizedImage(str(index), normalized_image)
+                        index += 1
+                else:
+                    print('No document found')
+        elif ch == ord('s'): # save image
+            if len(g_normalized_images) > 0:
+                for data in g_normalized_images:
+                    # cv2.imwrite('images/' + str(time.time()) + '.png', image)
+                    cv2.destroyWindow(data[0])
+                    data[1].save(str(time.time()) + '.png')
+                    print('Image saved')
+                    data[1].recycle()
+                    
                 g_normalized_images = []
                 index = 0
-                for result in g_results:
-                    x1 = result.x1
-                    y1 = result.y1
-                    x2 = result.x2
-                    y2 = result.y2
-                    x3 = result.x3
-                    y3 = result.y3
-                    x4 = result.x4
-                    y4 = result.y4
-                    
-                    normalized_image = scanner.normalizeBuffer(image, x1, y1, x2, y2, x3, y3, x4, y4)
-                    g_normalized_images.append((str(index), normalized_image))
-                    mat = showNormalizedImage(str(index), normalized_image)
-                    index += 1
-        elif ch == ord('s'): # save image
-            for data in g_normalized_images:
-                # cv2.imwrite('images/' + str(time.time()) + '.png', image)
-                cv2.destroyWindow(data[0])
-                data[1].save(str(time.time()) + '.png')
-                print('Image saved')
-                data[1].recycle()
+            else:
+                print('No image to save')
                 
         if image is not None:
             scanner.detectMatAsync(image)
@@ -69,9 +79,9 @@ def process_video(scanner):
                 
                 cv2.drawContours(image, [np.int0([(x1, y1), (x2, y2), (x3, y3), (x4, y4)])], 0, (0, 255, 0), 2)
             
-        cv2.putText(image, 'Press "n" to normalize image', (10, 30), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
-        cv2.putText(image, 'Press "s" to save image', (10, 60), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
-        cv2.putText(image, 'Press "ESC" to exit', (10, 90), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
+        cv2.putText(image, '1. Press "n" to normalize image', (10, 30), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
+        cv2.putText(image, '2. Press "s" to save image', (10, 60), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
+        cv2.putText(image, '3. Press "ESC" to exit', (10, 90), cv2.FONT_HERSHEY_SIMPLEX, 0.8, (0, 0, 255), 2)
         cv2.imshow('Document Scanner', image)
         
     for data in g_normalized_images:
